simetuc/requirements.txt
ase>=3.9.1 # pip
matplotlib>=1.5.1 # anaconda
numpy>=1.11.1 # anaconda
pytest>=3.0.3 # anaconda, only for tests
scipy>=0.18.1 # anaconda
tqdm>=4.8.4 # pip
PyYAML>=3.12 # anaconda
